pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bq-pubsub-job"
version = "0.1.0"
description = "BigQuery to Pub/Sub ETL job for GitHub Archive data"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-bigquery==3.11.4",
    "google-cloud-pubsub==2.18.4",
    "functions-framework==3.4.0",
    "python-dateutil==2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "ruff>=0.1.0",
]